+++ src/main/java/com/tunisair/service/EquipageService.java
@@ -1,35 +1,44 @@
-package com.tunisair.services;
+package com.tunisair.service;
 
 import com.tunisair.models.Equipage;
 import com.tunisair.repositories.EquipageRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import javassist.NotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class EquipageService {
+
     private final EquipageRepository equipageRepository;
 
-    @Autowired
     public EquipageService(EquipageRepository equipageRepository) {
         this.equipageRepository = equipageRepository;
     }
 
-    // Implement service methods for Equipage class
-    public List<Equipage> getAllEquipages() {
-        return equipageRepository.findAll();
+    public Equipage getEquipageById(Long id) throws NotFoundException {
+        return equipageRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Equipage not found with id: " + id));
     }
 
-    public Equipage getEquipageById(Long id) {
-        return equipageRepository.findById(id).orElse(null);
-    }
-
-    public Equipage saveEquipage(Equipage equipage) {
+    public Equipage createEquipage(Equipage equipage) {
         return equipageRepository.save(equipage);
     }
 
-    public void deleteEquipageById(Long id) {
+    public Equipage updateEquipage(Long id, Equipage equipage) throws NotFoundException {
+        Equipage existingEquipage = equipageRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Equipage not found with id: " + id));
+
+        // Update attributes of existingEquipage with the new values from equipage
+        existingEquipage.setPilotes(equipage.getPilotes());
+        existingEquipage.setStaffs(equipage.getStaffs());
+        existingEquipage.setVols(equipage.getVols());
+
+        return equipageRepository.save(existingEquipage);
+    }
+
+    public void deleteEquipage(Long id) throws NotFoundException {
+        if (!equipageRepository.existsById(id)) {
+            throw new NotFoundException("Equipage not found with id: " + id);
+        }
         equipageRepository.deleteById(id);
     }
 }
